Fixed a bug where next could advance to a removed item. Changed the way iterator is used. Added capacity argument to hashset_create().

+++ include/core/utils/hashset/hashset_itr.h
@@ -14,6 +14,7 @@
  *   limitations under the License.
  *
  *  Modified in 2022 by Edward A. Lee to conform to documentation standards.
+ *  Also, changed the logic to iterate using just hashset_iterator_next().
  */
 
 #ifndef HASHSET_ITR_H_
@@ -27,13 +28,23 @@ extern "C" {
 
 struct hashset_itr_st {
   hashset_t set;
-  size_t index;
+  int index;
 };
 
 typedef struct hashset_itr_st *hashset_itr_t;
 
 /**
  * @brief Create a hashset iterator.
+ * The caller should then iterate over the hashset as follows:
+ * ```
+ *   hashset_itr_t iterator = hashset_iterator(my_hashset);
+ *   while (hashset_iterator_next(iterator) >= 0) {
+ *     void* my_value = hashset_iterator_value(iterator);
+ *     ...
+ *   }
+ *   free(iterator);
+ * ```
+ * The caller must call free() on this iterator after using it. 
  */
 hashset_itr_t hashset_iterator(hashset_t set);
 
@@ -50,7 +61,7 @@ int hashset_iterator_has_next(hashset_itr_t itr);
 
 /**
  * @brief Advance to the next value in the hashset.
- * This returns a positive number (the current index) if there is a next item
+ * This returns a non-negative number (the current index) if there is a next item
  * and -1 otherwise.
  */
 int hashset_iterator_next(hashset_itr_t itr);
